# api/openapi.yaml
openapi: 3.0.3
info:
  title: go-httpbin
  version: 0.1.0
  description: HTTP request and response testing service with practical httpbin parity.
servers:
  - url: /
paths:
...
  /post:
    post:
      operationId: PostEcho
      requestBody:
        content:
          "*/*":
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Request echo.
  /put:
    put:
      operationId: PutEcho
      requestBody:
        content:
          "*/*":
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Request echo.
  /patch:
    patch:
      operationId: PatchEcho
      requestBody:
        content:
          "*/*":
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Request echo.
...
  /anything:
    get:
      operationId: AnythingGet
      responses:
        "200":
          description: Request echo.
    post:
      operationId: AnythingPost
      requestBody:
        content:
          "*/*":
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Request echo.
    put:
      operationId: AnythingPut
      requestBody:
        content:
          "*/*":
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Request echo.
    patch:
      operationId: AnythingPatch
      requestBody:
        content:
          "*/*":
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Request echo.
    delete:
      operationId: AnythingDelete
      responses:
        "200":
          description: Request echo.
  /anything/{path}:
    parameters:
      - $ref: "#/components/parameters/Path"
    get:
      operationId: AnythingPathGet
      responses:
        "200":
          description: Request echo.
    post:
      operationId: AnythingPathPost
      requestBody:
        content:
          "*/*":
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Request echo.
    put:
      operationId: AnythingPathPut
      requestBody:
        content:
          "*/*":
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Request echo.
    patch:
      operationId: AnythingPathPatch
      requestBody:
        content:
          "*/*":
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Request echo.
    delete:
      operationId: AnythingPathDelete
      responses:
        "200":
          description: Request echo.
...
  /status/{codes}:
    parameters:
      - $ref: "#/components/parameters/Codes"
    get:
      operationId: StatusGet
      responses:
        default:
          description: Requested status.
    post:
      operationId: StatusPost
      responses:
        default:
          description: Requested status.
    put:
      operationId: StatusPut
      responses:
        default:
          description: Requested status.
    delete:
      operationId: StatusDelete
      responses:
        default:
          description: Requested status.
...
  /cache/{seconds}:
    parameters:
      - $ref: "#/components/parameters/Seconds"
    get:
      operationId: CacheSeconds
      responses:
        "200":
          description: Cacheable response.
...
  /basic-auth/{user}/{password}:
    parameters:
      - $ref: "#/components/parameters/User"
      - $ref: "#/components/parameters/Password"
    get:
      operationId: BasicAuth
      responses:
        "200":
          description: Authenticated.
        "401":
          description: Unauthorized.
  /hidden-basic-auth/{user}/{password}:
    parameters:
      - $ref: "#/components/parameters/User"
      - $ref: "#/components/parameters/Password"
    get:
      operationId: HiddenBasicAuth
      responses:
        "200":
          description: Authenticated.
        "404":
          description: Hidden unauthorized response.
...
  /stream-bytes/{n}:
    parameters:
      - $ref: "#/components/parameters/N"
    get:
      operationId: StreamBytes
      responses:
        "200":
          description: Streamed random bytes.
...
  /base64/{value}:
    parameters:
      - $ref: "#/components/parameters/Value"
    get:
      operationId: Base64Decode
      responses:
        "200":
          description: Decoded base64.
...
  /cookies/set/{name}/{value}:
    parameters:
      - $ref: "#/components/parameters/Name"
      - $ref: "#/components/parameters/Value"
    get:
      operationId: SetCookie
      responses:
        "302":
          description: Redirect to cookie listing.
...
  /redirect/{n}:
    parameters:
      - $ref: "#/components/parameters/N"
    get:
      operationId: Redirect
      responses:
        "302":
          description: Redirect.
  /relative-redirect/{n}:
    parameters:
      - $ref: "#/components/parameters/N"
    get:
      operationId: RelativeRedirect
      responses:
        "302":
          description: Relative redirect.
  /absolute-redirect/{n}:
    parameters:
      - $ref: "#/components/parameters/N"
    get:
      operationId: AbsoluteRedirect
      responses:
        "302":
          description: Absolute redirect.
...
components:
  parameters:
    Codes:
      name: codes
      in: path
      required: true
      schema:
        type: string
...
    N:
      name: n
      in: path
      required: true
      schema:
        type: integer
    Name:
      name: name
      in: path
      required: true
      schema:
        type: string
...
    Password:
      name: password
      in: path
      required: true
      schema:
        type: string
    Path:
      name: path
      in: path
      required: true
      schema:
        type: string
    Seconds:
      name: seconds
      in: path
      required: true
      schema:
        type: integer
    User:
      name: user
      in: path
      required: true
      schema:
        type: string
    Value:
      name: value
      in: path
      required: true
      schema:
        type: string